+++ app/src/main/java/ru/vavtech/septemberworkout/list/WorkoutAdapter.java
@@ -1,0 +1,38 @@
+package ru.vavtech.septemberworkout.list;
+
+import android.support.annotation.NonNull;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import java.util.List;
+
+import ru.vavtech.septemberworkout.Model.Workout;
+import ru.vavtech.septemberworkout.Model.WorkoutList;
+import ru.vavtech.septemberworkout.R;
+
+public class WorkoutAdapter extends RecyclerView.Adapter<WorkoutViewHOlder> {
+    private List<Workout> workoutList = WorkoutList.getInstance().getWorkouts();
+
+    @NonNull
+    @Override
+    public WorkoutViewHOlder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
+        View itemView = LayoutInflater.from(viewGroup.getContext()).inflate(
+                R.layout.list_item,
+                viewGroup,
+                false
+        );
+        return new WorkoutViewHOlder(itemView);
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull WorkoutViewHOlder workoutViewHOlder, int index) {
+        workoutViewHOlder.bindView(workoutList.get(index));
+    }
+
+    @Override
+    public int getItemCount() {
+        return workoutList != null ? workoutList.size() : 0;
+    }
+}
